imlement file uploader

+++ controllers/admin.js
@@ -1,5 +1,8 @@
 // const mongodb = require('mongodb');
 const Product = require('../models/product');
+
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator/check');
 // const ObjectId = mongodb.ObjectId;//mongoDBでidを使用する場合はコンストラクターの引数にセットする必要がある
 
@@ -17,9 +20,25 @@ exports.getAddProduct = (req,res,next)=>{ //第一引数には相対パスをセ
 exports.postAddProduct =(req,res,next)=>{
     // console.log(req.body);//フォームからの入力値をparseするにはbody-parserが必要
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    if(!image){
+      return res.status(422).render('admin/edit-product',{
+        pageTitle:'Add Product',
+        path:'/admin/add-product',
+        editing:false,
+        product:{
+          title:title,
+          price:price,
+          description:description
+        },
+        hasError:true,
+        errorMessage:'Attached file is not an image.',
+        validationErrors:[]
+      });
+    }
+
     const errors = validationResult(req);
 
     if(!errors.isEmpty()){
@@ -29,7 +48,6 @@ exports.postAddProduct =(req,res,next)=>{
         editing:false,
         product:{
           title:title,
-          imageUrl:imageUrl,
           price:price,
           description:description
         },
@@ -38,6 +56,8 @@ exports.postAddProduct =(req,res,next)=>{
         validationErrors:errors.array()
       });
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({
       title:title,
@@ -115,7 +135,7 @@ exports.postEditProduct = (req,res,next)=>{
     
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDesc = req.body.description;
 
     const errors = validationResult(req);
@@ -127,7 +147,6 @@ exports.postEditProduct = (req,res,next)=>{
         editing:true,
         product:{
           title:updatedTitle,
-          imageUrl:updatedImageUrl,
           price:updatedPrice,
           description:updatedDesc,
           _id:prodId
@@ -145,7 +164,10 @@ exports.postEditProduct = (req,res,next)=>{
       product.title =updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if(image){
+        fileHelper.deleteFile(product.imageUrl);//既存のファイルを削除
+        product.imageUrl = image.path;
+      }
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
@@ -177,8 +199,14 @@ exports.getProducts = (req,res,next)=>{
 }
 exports.postDeleteProduct =(req,res,next)=>{
     const prodId = req.body.productId;
-    Product.deleteOne({_id:prodId,userId:req.user._id})
-    .then(() => {
+    Product.findById(prodId)
+      .then(product =>{
+      if(!product){
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);//既存のファイルを削除
+      return Product.deleteOne({_id:prodId,userId:req.user._id})
+    }).then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     }).catch(err => {
